initial prototype without fixing test cases
